refactoring methods for function calls
