+++ src/passes/loopFunctionaliser/whileLoopToFunction.ts
@@ -1,9 +1,10 @@
 import { DoWhileStatement, FunctionDefinition, WhileStatement } from 'solc-typed-ast';
 import { AST } from '../../ast/ast';
 import { ASTMapper } from '../../ast/mapper';
+import { createOuterCall } from '../../utils/functionGeneration';
 import {
   collectUnboundVariables,
-  createOuterCall,
+  createLoopCall,
   extractDoWhileToFunction,
   extractWhileToFunction,
 } from './utils';
@@ -27,7 +28,13 @@ export class WhileLoopToFunction extends ASTMapper {
       this.loopToContinueFunction,
       ast,
     );
-    const outerCall = createOuterCall(node, unboundVariables, functionDef, ast);
+    // const outerCall = createOuterCall(node, unboundVariables, functionDef, ast);
+    const outerCall = createOuterCall(
+      node,
+      [...unboundVariables.keys()],
+      createLoopCall(functionDef, [...unboundVariables.keys()], ast),
+      ast,
+    );
     ast.replaceNode(node, outerCall);
   }
